feat: 'games 앱 분리'
