+++ back/config/urls.py
@@ -23,5 +23,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('dj_rest_auth.urls')),
     path('api/v1/', include('movies.urls')),
+    path('api/v2/', include('games.urls')),
     path('accounts/signup/', include('dj_rest_auth.registration.urls')),
 ] + static(settings.MEDIA_URL, document_root = settings.MEDIA_ROOT)
